requirements.txt
discord.py
python-dotenv
GitPython
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
requests